central_system/requirements.txt
# AI/ML Dependencies
google-generativeai>=0.3.0
openai>=1.0.0
httpx>=0.25.0

# FastAPI Core
fastapi==0.104.1
uvicorn[standard]==0.24.0

# WebSocket Support
websockets==12.0

# Database
sqlalchemy==2.0.23
aiosqlite==0.19.0

# Utilities
python-dotenv==1.0.0
ujson==5.8.0
psutil==5.9.6

# Monitoring & Logging
structlog==23.1.0